requirements.txt
Flask>=2.3,<4
PyMySQL>=1.1,<2
python-dotenv>=1.0,<2
